+++ examples/callcontext/client.py
@@ -11,17 +11,15 @@ if sys.version_info < (3, 0):
     input = raw_input
 
 
+# custom proxy needed to get to annotation data, before Pyro 4.56
 class CustomAnnotationProxy(Pyro4.Proxy):
     def __init__(self, uri):
         super(CustomAnnotationProxy, self).__init__(uri)
         self._pyroHmacKey = b"secr3t_k3y"
 
-    # override the method that adds annotations and add our own,
-    # but be sure to call the base class method.
+    # override the method that adds annotations and add our own
     def _pyroAnnotations(self):
-        annotations = super(CustomAnnotationProxy, self)._pyroAnnotations()
-        annotations["XYZZ"] = b"Hello, I am a custom annotation from the proxy!"
-        return annotations
+        return {"XYZZ": b"Hello, I am a custom annotation from the proxy!"}
 
     def _pyroResponseAnnotations(self, annotations, msgtype):
         print("    Got response (msgtype=%d). Annotations:" % msgtype)
@@ -36,6 +34,8 @@ class CustomAnnotationProxy(Pyro4.Proxy):
 
 
 uri = input("Enter the URI of the server object: ")
+
+print("\n------- (older) method to get annotations via callback on custom proxy... -----\n")
 with CustomAnnotationProxy(uri) as proxy:
     print("Sending a few messages using one proxy...")
     for i in range(4):
@@ -58,4 +58,23 @@ with CustomAnnotationProxy(uri) as proxy:
     print("Finally, sending a oneway message...")
     proxy.oneway("hello-ONEWAY")
 
-print("See the console output on the server for more results.")
+
+print("\n------- get annotations via normal proxy and the call context... -----\n")
+input("press enter:")
+# the code below works as of Pyro 4.56.
+with Pyro4.Proxy(uri) as proxy:
+    proxy._pyroHmacKey = b"secr3t_k3y"
+    print("normal call")
+
+    Pyro4.current_context.annotations = {"XYZZ": b"custom annotation from client via new way(1)"}
+    result = proxy.echo("hi there - new method of annotation access in client")
+    print("Annotations in response were: ", Pyro4.current_context.response_annotations)
+
+    print("\noneway call")
+    Pyro4.current_context.annotations = {"XYZZ": b"custom annotation from client via new way(2)"}
+    proxy.oneway("hi there ONEWAY - new method of annotation access in client")
+    print("Annotations in response were: ", Pyro4.current_context.response_annotations)
+    print("   (should be empty because oneway!)")
+
+
+print("\nSee the console output on the server for more results.")