+++ tests/PyroTests/test_daemon.py
@@ -90,7 +90,7 @@ class DaemonTests(unittest.TestCase):
             self.assertIn(Pyro4.constants.DAEMON_NAME, d.objectsById)
             self.assertEqual("PYRO:" + Pyro4.constants.DAEMON_NAME + "@" + d.locationStr, str(d.uriFor(Pyro4.constants.DAEMON_NAME)))
             # check the string representations
-            expected = "<Pyro4.core.Daemon at 0x%x; %s; 1 objects>" % (id(d), d.locationStr)
+            expected = "<Pyro4.core.Daemon at 0x%x; %s - %s; 1 objects>" % (id(d), d.locationStr, Pyro4.socketutil.family_str(d.sock))
             self.assertEqual(expected, str(d))
             self.assertEqual(expected, unicode(d))
             self.assertEqual(expected, repr(d))
@@ -118,7 +118,7 @@ class DaemonTests(unittest.TestCase):
             self.assertEqual(locationstr, d.locationStr)
             self.assertEqual("PYRO:" + Pyro4.constants.DAEMON_NAME + "@" + locationstr, str(d.uriFor(Pyro4.constants.DAEMON_NAME)))
             # check the string representations
-            expected = "<Pyro4.core.Daemon at 0x%x; %s; 1 objects>" % (id(d), locationstr)
+            expected = "<Pyro4.core.Daemon at 0x%x; %s - Unix; 1 objects>" % (id(d), locationstr)
             self.assertEqual(expected, str(d))
             self.assertEqual(SOCKNAME, d.sock.getsockname())
             self.assertEqual(socket.AF_UNIX, d.sock.family)
@@ -131,7 +131,7 @@ class DaemonTests(unittest.TestCase):
             self.assertEqual(locationstr, d.locationStr)
             self.assertEqual("PYRO:" + Pyro4.constants.DAEMON_NAME + "@" + locationstr, str(d.uriFor(Pyro4.constants.DAEMON_NAME)))
             # check the string representations
-            expected = "<Pyro4.core.Daemon at 0x%x; %s; 1 objects>" % (id(d), locationstr)
+            expected = "<Pyro4.core.Daemon at 0x%x; %s - Unix; 1 objects>" % (id(d), locationstr)
             self.assertEqual(expected, str(d))
             sn_bytes = tobytes(SOCKNAME)
             self.assertEqual(sn_bytes, d.sock.getsockname())
@@ -412,9 +412,7 @@ class DaemonTests(unittest.TestCase):
             def validateHandshake(self, conn, data):
                 return ["sure", "have", "fun"]
             def annotations(self):
-                ann = super(CustomHandshakeDaemon, self).annotations()
-                ann["XYZZ"] = b"custom annotation set by daemon"
-                return ann
+                return {"XYZZ": b"custom annotation set by daemon"}
         with CustomHandshakeDaemon(port=0) as d:
             corr_id = uuid.uuid4()
             self.sendHandshakeMessage(conn, correlation_id=corr_id)